modified: transfer template to tests
