+++ SpaceBattle.Tests/AdapterGenerationTests.cs
@@ -1,4 +1,6 @@
-﻿public class AdapterGenerationTests
+﻿using Scriban;
+
+public class AdapterGenerationTests
 {
     public AdapterGenerationTests()
     {
@@ -8,6 +10,29 @@
             "Scopes.Current.Set",
             IoC.Resolve<object>("Scopes.New", IoC.Resolve<object>("Scopes.Root"))
         ).Execute();
+
+        var template = Template.Parse(@"using System;
+public class {{ structure_name }}Adapter : {{ structure_name }}
+{
+    private IUObject _obj;
+    public {{ structure_name }}Adapter(IUObject obj) => _obj = obj;{{for property in (properties)}}
+    public {{property.property_type.name}} {{property.name}}
+    {
+    {{if property.can_read}}
+    get
+    {
+        return ({{property.property_type.name}})_obj.GetProperty(""{{property.name}}"");
+    }{{end}}
+    {{if property.can_write}}
+    set
+    {
+        _obj.SetProperty(""{{property.name}}"", value);
+    }{{end}}
+    }
+{{end}}
+}");
+        IoC.Resolve<Hwdtech.ICommand>("IoC.Register", "AdapterSample", (object[] args) => (Template)template).Execute();
+
     }
 
     [Fact]
